refactor: split corporate-pieces.js into section modules + barrel
The 2305-line set-piece catalog became unwieldy. Split it into 9 themed
modules under data/biomes/corporate-pieces/ (defense, traps, puzzles, filler,
scaled, scattered, atomics, scenarios, backbone), with corporate-pieces.js
now a thin barrel that re-exports everything and assembles SET_PIECES.

Public surface is identical — no consumer (corporate.js, the network files,
tests) changed. Pure code movement: every set-piece definition is byte-identical
(verified by line-level diff). make check green (1154 tests).
